+++ lib/rover.rb
@@ -11,12 +11,12 @@ module Rover
   class << self
     def read_csv(path, types: nil, **options)
       require "csv"
-      csv_to_df(CSV.read(path, **csv_options(options)), types: types)
+      csv_to_df(CSV.read(path, **csv_options(options)), types: types, headers: options[:headers])
     end
 
     def parse_csv(str, types: nil, **options)
       require "csv"
-      csv_to_df(CSV.parse(str, **csv_options(options)), types: types)
+      csv_to_df(CSV.parse(str, **csv_options(options)), types: types, headers: options[:headers])
     end
 
     private
@@ -28,7 +28,11 @@ module Rover
       options
     end
 
-    def csv_to_df(table, types: nil)
+    def csv_to_df(table, types: nil, headers: nil)
+      if headers && headers.size < table.headers.size
+        raise ArgumentError, "Expected #{table.headers.size} headers, got #{headers.size}"
+      end
+
       table.by_col!
       data = {}
       table.each do |k, v|